requirements.txt
absl-py==2.1.0
aiohappyeyeballs==2.3.4
aiohttp==3.10.1
aiosignal==1.3.1
annotated-types==0.7.0
anthropic==0.32.0
anyascii==0.3.2
anyio==4.4.0
asttokens==2.4.1
async-timeout==4.0.3
attrs==24.1.0
audioread==3.0.1
av==12.3.0
Babel==2.15.0
bangla==0.0.2
blinker==1.8.2
blis==0.7.11
bnnumerizer==0.0.2
bnunicodenormalizer==0.1.7
catalogue==2.0.10
certifi==2024.7.4
cffi==1.16.0
charset-normalizer==3.3.2
click==8.1.7
cloudpathlib==0.18.1
coloredlogs==15.0.1
comm==0.2.2
confection==0.1.5
contourpy==1.2.1
coqpit==0.0.17
ctranslate2==4.3.1
cycler==0.12.1
cymem==2.0.8
Cython==3.0.11
dataclasses-json==0.6.7
dateparser==1.1.8
debugpy==1.8.5
decorator==5.1.1
defusedxml==0.7.1
distro==1.9.0
dlib==19.24.5
docopt==0.6.2
einops==0.8.0
elevenlabs==1.6.1
encodec==0.1.1
exceptiongroup==1.2.2
executing==2.0.1
face-recognition==1.3.0
face-recognition-models==0.3.0
fastapi==0.112.0
faster-whisper==1.0.3
filelock==3.15.4
Flask==3.0.3
flatbuffers==24.3.25
fonttools==4.53.1
frozenlist==1.4.1
fsspec==2024.6.1
g2pkk==0.1.2
greenlet==3.0.3
groq==0.9.0
grpcio==1.65.4
gruut==2.2.3
gruut-ipa==0.13.0
gruut-lang-fr==2.0.2
gruut_lang_de==2.0.1
gruut_lang_en==2.0.1
gruut_lang_es==2.0.1
h11==0.14.0
hangul-romanize==0.1.0
httpcore==1.0.5
httpx==0.27.0
httpx-sse==0.4.0
huggingface-hub==0.24.5
humanfriendly==10.0
idna==3.7
inflect==7.3.1
ipykernel==6.29.5
ipython==8.26.0
itsdangerous==2.2.0
jamo==0.4.1
jedi==0.19.1
jieba==0.42.1
Jinja2==3.1.4
jiter==0.5.0
joblib==1.4.2
jsonlines==1.2.0
jsonpatch==1.33
jsonpointer==3.0.0
jupyter_client==8.6.2
jupyter_core==5.7.2
kaldiio==2.18.0
kiwisolver==1.4.5
langchain==0.2.12
langchain-anthropic==0.1.22
langchain-community==0.2.11
langchain-core==0.2.28
langchain-groq==0.1.9
langchain-mistralai==0.1.12
langchain-openai==0.1.20
langchain-text-splitters==0.2.2
langcodes==3.4.0
langgraph==0.1.19
langsmith==0.1.96
language_data==1.2.0
lazy_loader==0.4
librosa==0.10.0
llvmlite==0.43.0
marisa-trie==1.2.0
Markdown==3.6
markdown-it-py==3.0.0
MarkupSafe==2.1.5
marshmallow==3.21.3
matplotlib==3.8.4
matplotlib-inline==0.1.7
mdurl==0.1.2
more-itertools==10.3.0
MouseInfo==0.1.3
mpmath==1.3.0
msgpack==1.0.8
multidict==6.0.5
murmurhash==1.0.10
mypy-extensions==1.0.0
nest-asyncio==1.6.0
networkx==2.8.8
nltk==3.8.1
num2words==0.5.13
numba==0.60.0
numpy==1.22.0
nvidia-cublas-cu12==12.1.3.1
nvidia-cuda-cupti-cu12==12.1.105
nvidia-cuda-nvrtc-cu12==12.1.105
nvidia-cuda-runtime-cu12==12.1.105
nvidia-cudnn-cu12==9.1.0.70
nvidia-cufft-cu12==11.0.2.54
nvidia-curand-cu12==10.3.2.106
nvidia-cusolver-cu12==11.4.5.107
nvidia-cusparse-cu12==12.1.0.106
nvidia-nccl-cu12==2.20.5
nvidia-nvjitlink-cu12==12.6.20
nvidia-nvtx-cu12==12.1.105
ollama==0.3.1
onnxruntime==1.18.1
openai==1.39.0
opencv-python==4.10.0.84
orjson==3.10.6
packaging==24.1
pandas==1.5.3
parso==0.8.4
pexpect==4.9.0
pillow==10.4.0
platformdirs==4.2.2
pooch==1.8.2
preshed==3.0.9
prompt_toolkit==3.0.47
protobuf==4.25.4
psutil==6.0.0
ptyprocess==0.7.0
pure_eval==0.2.3
PyAudio==0.2.14
PyAutoGUI==0.9.54
pycparser==2.22
pydantic==2.8.2
pydantic_core==2.20.1
pydub==0.25.1
PyGetWindow==0.0.9
Pygments==2.18.0
PyMsgBox==1.0.9
pynndescent==0.5.13
pyparsing==3.1.2
pyperclip==1.9.0
pypinyin==0.52.0
PyRect==0.2.0
pyrubberband==0.3.0
pysbd==0.3.4
PyScreeze==0.1.30
PySoundFile==0.9.0.post1
python-crfsuite==0.9.10
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-mpv==1.0.7
python3-xlib==0.15
pytweening==1.2.0
pytz==2024.1
PyYAML==6.0.1
pyzmq==26.1.0
regex==2024.7.24
requests==2.32.3
rich==13.7.1
safetensors==0.4.4
scikit-learn==1.5.1
scipy==1.11.4
shellingham==1.5.4
silero-vad==5.1
six==1.16.0
smart-open==7.0.4
sniffio==1.3.1
sounddevice==0.4.7
soundfile==0.12.1
soxr==0.4.0
spacy==3.7.5
spacy-legacy==3.0.12
spacy-loggers==1.0.5
SpeechRecognition==3.10.4
SQLAlchemy==2.0.31
srsly==2.4.8
stack-data==0.6.3
starlette==0.37.2
SudachiDict-core==20240716
SudachiPy==0.6.8
sympy==1.13.1
tenacity==8.5.0
tensorboard==2.17.0
tensorboard-data-server==0.7.2
thinc==8.2.5
threadpoolctl==3.5.0
tiktoken==0.7.0
tokenizers==0.19.1
torch==2.4.0
torchaudio==2.4.0
tornado==6.4.1
tqdm==4.66.5
trainer==0.0.36
traitlets==5.14.3
transformers==4.43.4
triton==3.0.0
TTS==0.22.0
typeguard==4.3.0
typer==0.12.3
typing-inspect==0.9.0
typing_extensions==4.12.2
tzlocal==5.2
umap-learn==0.5.6
Unidecode==1.3.8
urllib3==2.2.2
uvicorn==0.30.5
wasabi==1.1.3
wcwidth==0.2.13
weasel==0.4.1
websockets==12.0
Werkzeug==3.0.3
wespeaker==0.0.0
wrapt==1.16.0
yarl==1.9.4
youtube-search==2.1.2
